Move extensions validation to apps
